+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardStatus.kt
@@ -12,5 +12,6 @@ sealed class CardStatus {
     class VaultSuccess(val result: CardVaultResult) : CardStatus()
     class VaultCanceled(val setupTokenId: String?) : CardStatus()
 
-    object NoResult : CardStatus()
+    class UnknownError(val error: Throwable) : CardStatus()
+    data object NoResult : CardStatus()
 }