+++ CorePayments/src/main/java/com/paypal/android/corepayments/BrowserSwitchRequestCodes.kt
@@ -1,0 +1,9 @@
+package com.paypal.android.corepayments
+
+object BrowserSwitchRequestCodes {
+    const val CARD_APPROVE_ORDER = 1
+    const val CARD_VAULT = 2
+
+    const val PAYPAL_CHECKOUT = 3
+    const val PAYPAL_VAULT = 4
+}